requirements.txt
absl-py==0.9.0
astunparse==1.6.3
attrs==20.2.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
coverage==5.3
gast==0.3.3
google-auth==1.14.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.29.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.0
iniconfig==1.1.1
Keras-Preprocessing==1.1.2
Markdown==3.2.2
mxnet==1.6.0
numpy==1.18.4
oauthlib==3.1.0
opt-einsum==3.2.1
packaging==20.4
Pillow==7.1.2
pluggy==0.13.1
protobuf==3.11.3
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyparsing==2.4.7
pytest==6.1.1
pytest-cov==2.10.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
toml==0.10.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
